// src/lib/scanner-types.ts
export type PinStatus = "pending" | "scanning" | "finished" | "expired";
export type PinResult = "none" | "clean" | "warning" | "suspicious" | "cheating";

export type ScanDisplayTotals = {
  detections: number;
  warnings: number;
  suspicious: number;
  integrity: number;
  bamList: number;
  bypassList: number;
  powershellHistory: number;
  recordingSoftware: number;
  prefetchList: number;
  processList: number;
  adminLogs: number;
  browserList: number;
  usbList: number;
};

export interface CustomPattern {
  process: string;
  name: string;
  value: string;
  severity: string;
}
// ...
export interface PinDoc {
  _id?: string;
  pin: string;
  name?: string;
  game: string;
  ownerId: string;
  ownerName: string;
  status: PinStatus;
  used: boolean;
  result: PinResult;
  visibility: "private" | "public";
  enterpriseId?: string | null;
  disabled?: boolean;
  expiresAt?: Date | null;
  userPatterns?: CustomPattern[];
  progress?: number;
  scanMessage?: string;
  createdAt: Date;
  finishedAt?: Date | null;
  scanningStartedAt?: Date | null;
}